rename general_internal to basic
