+++ tobac/utils/internal/__init__.py
@@ -1,1 +1,1 @@
-from .general_internal import *
+from .basic import *
